update config file structure
