+++ pg_data_etl/__init__.py
@@ -1,6 +1,7 @@
-from .configuration_manager import configurations as connections
-
-
 import pg_data_etl.database.helpers as helpers
 import pg_data_etl.database.actions as actions
 from .database.Database import Database
+
+from .settings.read_config_file import configurations as _configurations
+
+config = _configurations()
